Change how delta tuples are processed. Previously, a newly-derived tuple was immediately inserted into/deleted from a table, and then inserted into a TupleBuf to be routed. This yielded the wrong behavior in the presence of refcounts, however: e.g.,
  p(A) :- q1(A), q2(A);

If we have q1(1) and q2(1) installed in the same fix point, the
previous behavior would bump the refcount for p(1) twice rather
than once.

Instead, insert newly-derived tuples into a TupleBuf for future
routing, but don't immediately modify any tables: we only modify
the table just prior to routing the tuple. This fixes the problem
described above: the first routed tuple doesn't join with anything,
so we end up only incrementing the refcount of p(1) once.

Add a regression test for this problem.